Split server crate into mock_server and server removing cli dependency on mdbx
