+++ rust/cli/Cargo.toml
@@ -15,12 +15,12 @@ call_server = { workspace = true }
 test_runner = { workspace = true }
 
 anyhow = { workspace = true }
-chain_server = { workspace = true }
 clap = { workspace = true, features = ["derive", "string"] }
 colored = { workspace = true }
 flate2 = { workspace = true }
 hex = { workspace = true }
 lazy_static = { workspace = true }
+mock_chain_server = { workspace = true }
 reqwest = { workspace = true, features = ["json"] }
 risc0-zkp = { workspace = true }
 serde = { workspace = true }
