+++ rust/services/call/host/Cargo.toml
@@ -15,12 +15,12 @@ call_guest_wrapper = { workspace = true }
 chain = { workspace = true }
 chain_client = { workspace = true, features = ["rpc"] }
 chain_common = { workspace = true }
-chain_server = { workspace = true }
 derive-new = { workspace = true }
 dotenvy = { workspace = true }
 ethers-core = { workspace = true }
 host_utils = { workspace = true }
 lazy_static = { workspace = true }
+mock_chain_server = { workspace = true }
 mpt = { workspace = true }
 provider = { workspace = true }
 reqwest = { workspace = true }
