+++ rust/services/chain/common/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2021"
 [dependencies]
 alloy-primitives = { workspace = true }
 block_trie = { workspace = true }
+mpt = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
