+++ rust/test_runner/src/cheatcode_inspector.rs
@@ -10,12 +10,12 @@ use call_engine::{
 use call_host::host::{config::HostConfig, get_block_number, Host};
 use chain::TEST_CHAIN_ID;
 use chain_client::RpcClient as RpcChainProofClient;
-use chain_server::server::{ChainProofServerMock, EMPTY_PROOF_RESPONSE};
 use foundry_config::RpcEndpoints;
 use foundry_evm::revm::{
     interpreter::{CallInputs, CallOutcome},
     Database, EvmContext, Inspector,
 };
+use mock_chain_server::{ChainProofServerMock, EMPTY_PROOF_RESPONSE};
 use provider::CachedMultiProvider;
 use serde_json::json;
